refactor(security): split oversized remediation modules
