+++ crates/tandem-runtime/src/mcp.rs
@@ -2,3 +2,4 @@ include!("mcp_parts/part01.rs");
 include!("mcp_parts/part02.rs");
 include!("mcp_parts/part04.rs");
 include!("mcp_parts/part03.rs");
+include!("mcp_parts/part05.rs");
