+++ crates/tandem-server/src/http/config_providers.rs
@@ -17,3 +17,4 @@ include!("config_providers_parts/part02.rs");
 include!("config_providers_parts/part03.rs");
 include!("config_providers_parts/part04.rs");
 include!("config_providers_parts/part05.rs");
+include!("config_providers_parts/part06.rs");
