+++ crates/tandem-server/src/pack_manager_tests.rs
@@ -1,3 +1,6 @@
+// Copyright (c) 2026 Frumu LTD
+// Licensed under the Business Source License 1.1
+
 use super::*;
 use ed25519_dalek::{Signer, SigningKey};
 use std::io::Write;
